apps/api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
drf-spectacular>=0.28,<1.0
psycopg[binary]>=3.2,<4.0
reportlab>=4.2,<5.0
Pillow>=10.0,<12.0
